refactor add functionality into own class
